+++ LiveSketch/SKTFoundationExtras.m
@@ -7,7 +7,7 @@
 @implementation NSObject (SKTPerformExtras)
 
 - (void)performSelector:(SEL)sel withEachObjectInArray:(NSArray *)array {
-    unsigned i, c = [array count];
+    NSUInteger i, c = [array count];
     for (i=0; i<c; i++) {
         [self performSelector:sel withObject:[array objectAtIndex:i]];
     }
